moved eventListener and querySelector from main.js to uiManager.js
